Add basic support for numeric range types
The syntax of a range is `(OP N)`, where OP is one of <, <=, >, or >=,
and N is a numeric literal. The parentheses are mandatory. The parser
refers to these as a HalfRangeType.

The rules for them are as follows:

Assigning to a range from a number or an enum is not allowed. In the
future, enums may be supported directly, while numbers will have to be
done through control flow analysis.

Assigning to a range from a number literal (including enum literals) is
allowed if the literal is within the range. Assigning from BigInt
literals is not supported as the compiler only treats them as strings.

Assigning from a range to another is only allowed if the source range is
contained entirely within the target range. For example, assigning from
(>= 0) to (> 0) is not allowed, because 0 would be valid in the source
but not the target. Assigning from (> 0) to (> 1) is not allowed, but
the reverse is.

Assigning from a range to a generic number is always allowed. Assigning
to an enum is not.

Internally, the compiler represents all ranges as "full" ranges, e.g.
[0, 1) in interval notation. In the future this will be used in union
and intersection types to "merge" ranges. For example, the union of
(>= 0) and (< 1) should be [0, 1), while the intersection should be
`never`. These should still work normally, but there isn't any special
support for them that would support such narrowing.

Right now, these "full" ranges are not actually supported in the syntax,
nor is this planned.

Since this is a first draft, there's not yet any support for language
services, and there's very limited diagnostic information available.

In the future, control flow analysis will be done to automatically widen
and narrow numbers and ranges as necessary.