WIP: Pass class static type to interface base type
This commit attempts to make progress towards #5863. It is not a
full solution, but should be enough for library writers to use a
polymorphic this for static members.

Consider the following:
```typescript
  interface StaticShape {
    new(): {};
    poly_prop: this;
  }

  // ...

  class Derived extends (X as StaticShape) {

  }
```

One difficultly with attempting to implement #5863, is how to
make it _clear_ which `this` the `this` type refers to. This
sidesteps the issue, by making it *explicit* which side the
`this` type falls on.