Remove usage of non-standardised `Reflect.decorate`.
`Reflect.decorate` is not on a standards track and has no active proposal being championed in TC39.

We have no idea if `Reflect.decorate` will get added to the language, even if it does, we have no guarantee it will be the implementation we are assuming it to be.

By having this condition in the code, we are likely preventing `Reflect.decorate` from ever existing in the language, just like `Array.prototype.flatten` had to get a different name because MooTools would attempt to use it if it existed and it was not the implementation MooTools assumed it would be.