Use @definitelytyped/types-registry for ATA
Fall back to types-registry if `npm install
@definitelytyped/types-registry` doesn't work.

Notes:

1. This currently requires you to be authenticated with npm. After the
Github package registry leaves beta, that will no longer be true.
2. The error handling was incorrect -- it double-handled an exception and
ignored the return code of execSyncAndLog. I corrected it so that it
logs if either npm or GHPR fails.
3. I haven't added tests for this change yet. I see some references to
types-registry in the tests so I'll see if that's the right place.