Backward-compatible accessor emit
Accessors are now only emitted in d.ts when useDefineForClassFields is set.
Otherwise, properties that originated as accessors are prefixed with
a JSDoc tag: `/**@accessor*/`

The checker now understands `/**@accessor*/` tags for the purpose of
error reporting; if --useDefineForClassFields is provided, an ambient
base property with `/**@accessor*/` will be treated as an accessor, not
as a property.