JSDoc: Resolve return type of construct signature
Pre 3.7, this was handled by the weird try-again catchall case. Now we need to explicitly
handle the case of value that has a constructable type:

```ts
declare var Image: { new (): HTMLImageElement }
```

This is not a commonly written construction, but the DOM uses it for
Image, so that means it's widely consumed. In Typescript, the reference
must be `typeof Image`, but in Javascript, JSDoc references should allow
`Image` as well.