Require all parameters in JS
This is an experiment in fixing #31888. This is the simplest experiment,
which requires all parameters to be provided in JS, same as TS. If this
doesn't break the user tests that are written in JS too badly, then I
won't proceed to the more complex experiments, which are to require all
paremeters for contextually typed functions, or to require all
parameters when `"noImplicitAny": true`.