Speed up getAccessibleFileSystemEntries
While writing a benchmark, I happened to notice that (a) the natural
win32 APIs for this task include "isDirectory" in the result and (b)
readDirSync will return entries, rather than strings, if you pass
withFileTypes.  It turns out this version, which allows us to dispense
with path math and stat calls is ~30X faster (enumerating a directory
with 600 children).