Make `Symbol#valueDeclaration` optional in public API
`compiler/types.ts` uses some non-nullable typings to avoid frequent
assertions in the TS compiler. This causes a loss of type information
and makes using typescript's public API dangerous since an external
consumer may assume that a certain type is defined when it is not.
(See #24706 for more details).

One such case is `Symbol#valueDeclaration`, which this PR provides a
public API typing fix for. This is implemented by adding an additional
processing step to the LGK compiler that overwrites definitions with the
`@optional` JSDoc tag to include the `?` token in their signature.
Specifically, `/** @optional */ valueDeclaration: Declaration` ->
`valueDeclaration?: Declaration`. The relevant change in the public API
definitions is at lib/typescript.d.ts:2285.

Part of #24706