Fix instantiated check for imports
Currently, the instantiated check for namespaces assumes that an
`import` that doesn't use the `export import` syntax is uninstantiated.
This is not correct when the import is exported.

This PR searchs siblings and descendants for exports of the symbol bound
by the import.

This is less efficient. I'm thinking about how to improve that.

Fixes #35385